requirements.txt
googletrans==2.4.0
python-telegram-bot==12.7